client/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
DateTime==5.5
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
numpy==2.1.1
openai==1.44.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==74.1.2
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uuid==1.30
zope.interface==7.0.3
